ml_service/requirements.txt
tensorflow>=2.15.0
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
joblib>=1.3.0
